requirements.txt
flask
python-dotenv
psycopg2
psycopg2-binary
uuid